Hack: write raw acceleration data
As a (hopefully temporary) quick solution to get some testing data,
accumulate all raw acceleration data in the sensor and then write it as
a JSON in stopListening() (i.e. after you stop a recortding and
disconnect from the device) to the application directory.  That makes it
possible (if awkward) to get from the device.